% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{auto_maxer}
\alias{auto_maxer}
\title{Function to select n_max earthquakes automatically}
\usage{
auto_maxer(data)
}
\arguments{
\item{data}{A data.frame of data which has been passed from
\code{geom_timeline_label} to \code{stat_timeline_label}.}
}
\value{
A data.frame of data with a new n_max column
}
\description{
If a value for n_max earthquakes is not supplied by the user, this
function calculates a reasonable n_max value and creates a new column
containing this new n_max value.
This n_max value is equal to half the number of rows in the data
inputted. If this is not an integer, it is rounded down.
A message will be displayed indicating what value n_max was set to.
An error is thrown if n_max is calculated to be less than 1 as this
will be impossible to plot.
}
\examples{
  \dontrun{auto_maxer(tail(eq,20))}
  \dontrun{auto_maxer(tail(eq,19))}

}
