% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{clean_local_location}
\alias{clean_local_location}
\title{Clean up an earthquake location string}
\usage{
clean_local_location(local_loc_str)
}
\arguments{
\item{local_loc_str}{A character string giving the locations 
separated by commas.}
}
\value{
This function returns a string containing the cleaned up 
 location.
}
\description{
This function cleans up the local location of an earthquake.
If there are multiple locations separated by commas, each 
location name is cleaned and the locations are recombined into a 
string. Cleaning involves removing extra white space, 
converting to title case, and removing location names that are 
only directions, e.g., "e", "east of", etc. This function 
is used by eq_clean_data().
}
\examples{
clean_local_location("CUZCO,COLLAO,LIMA")

}
