% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{clipper}
\alias{clipper}
\title{A data clipper function for \code{stat_timeline}}
\usage{
clipper(data)
}
\arguments{
\item{data}{A data.frame to be plotted in ggplot2 which is passed to
\code{geom_timeline()} and subsequently \code{stat_timeline()}.}
}
\value{
A data.frame filtered to contain only the observations lying
within the \code{xmin} and \code{xmax} bounds specified.
}
\description{
This function "clips" the data that has been passed to the timeline
stat. All observations that occur before the specified xmin or
minimum date are filtered out, as are all observations that occur
after the specified xmax or maximum date. The "clipped" data frame is
then returned and used by the timeline geom for plotting.
}
