% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{clipper_plus}
\alias{clipper_plus}
\title{A data clipper function adapted for \code{stat_timeline_label}}
\usage{
clipper_plus(data)
}
\arguments{
\item{data}{A data.frame of data which has been passed from
\code{geom_timeline_label} to \code{stat_timeline_label}.}
}
\value{
A data.frame filtered to contain only the observations lying
within the \code{xmin} and \code{xmax} bounds specified AND
filtered to contain only n_max earthquakes to plot.
}
\description{
This function "clips" the data that has been passed to the timeline
label stat. All observations that occur before the specified xmin or
minimum date are filtered out, as are all observations that occur
after the specified xmax or maximum date.
The data is then filtered so that the n_max largest observations will
be flagged with the logical value TRUE. This indicates to the geom
that these observations should be plotted with a label.
}
\examples{
  \dontrun{clipper_plus(data.frame(x=c(4,7,12),
                                   xmin=c(6,6,6),
                                   xmax=c(15,15,15),
                                   txt=c("red","yellow","blue"),
                                   max_by=c(1,2,3)))}

}
