% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_data_cleaning.R
\name{country_finder}
\alias{country_finder}
\title{Function to classify lon/lat points as countries}
\usage{
country_finder(data)
}
\arguments{
\item{data}{A tibble passed to the function through the pipe chain
within the \code{eq_raw_cleaner} function.}
}
\value{
A tibble with a Country column
}
\description{
This function was necessary as the author of this code did not have
access to the exact data set specified in the instructions.
See README for details.
The data set used did not have a Country column. This function works
around that problem by creating a Country column by classifying each
entry (by latitude and longitude) as a country as defined in the
package \code{rnaturalearth}, wherever possible.
}
\examples{
  \dontrun{country_finder(tail(eq,50))}
  \dontrun{country_finder(data.frame(Latitude = 23, Longitude=33))}

}
