% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_data_cleaning.R
\name{date_paster}
\alias{date_paster}
\title{A function to paste together dates}
\usage{
date_paster(y, m = 1, d = 1)
}
\arguments{
\item{y}{A year number as a character string}

\item{m}{A month number (1-12) as a character string. Defaults to 1
(January)}

\item{d}{A day number as a character string. Defaults to 1 (the 1st).}
}
\value{
A date of class Date in y-m-d format
}
\description{
A function to paste together dates
}
\examples{
\dontrun{date_paster(y=2022)}
\dontrun{date_paster(y=2020, m=8)}

}
