% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_data_cleaning.R
\name{date_sorter}
\alias{date_sorter}
\title{A function to sort dates}
\usage{
date_sorter(yvec, mvec, dvec)
}
\arguments{
\item{yvec}{A numeric vector of years}

\item{mvec}{A numeric vector of months}

\item{dvec}{A numeric vector of days}
}
\value{
A column of type Date containing dates in the format y-m-d.
All dates in this column are correctly formatted.
}
\description{
This function uses the helper functions \code{bc_date()},
\code{year_pad()} and \code{date_paster()} to sort vectors of years,
months, and days (the columns Year, Mo, and Dy in the NOAA data).
It first checks for BC dates and handles those with the function
\code{bc_date()}. It then handles AD dates. First, it uses the
\code{year_pad()} function to handle 2- and 3- digit years, then uses
the \code{date_paster()} function on 4-digit years.
}
\examples{
  \dontrun{date_sorter(yvec=c("-200","9","1982"),
                       mvec=c("2","7","11"),
                       dvec=c("5","10","22")}

}
