% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataframe.R
\name{download_NOAA}
\alias{download_NOAA}
\title{Function to download NOAA data (live download not yet working)}
\usage{
download_NOAA()
}
\value{
raw data from the NOAA website
}
\description{
Function to download NOAA data (live download not yet working)
}
\examples{
\dontrun{download_NOAA()}
}
