% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_raw.R
\docType{data}
\name{earthquake_raw}
\alias{earthquake_raw}
\title{Raw NOAA dataset}
\format{A data frame with 5956 rows and 47 variables}
\usage{
data(earthquake_raw)
}
\description{
A data.frame with a raw NOAA dataset loaded into R. This is the data.frame that we use in all of our examples.
A description of the field names can be found at the NOAA website: \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.
}
\keyword{datasets}
