% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\docType{data}
\name{EarthquakeData}
\alias{EarthquakeData}
\title{EarthquakeData}
\format{
A data frame with 6,243 rows and 39 columns:
\describe{
  \item{Search parameter}{Search Parameter for database (not important)}
  \item{Year}{The year of the earthquake}
  \item{Mo}{The month of the earthquake}
  \item{Dy}{The day of the earthquake}
  \item{Hr}{The hour of the earthquake}
  \item{Mn}{The minute of the earthquake}
  \item{Sec}{The second of the earthquake}
  \item{Tsu}{Related tsunami}
  \item{Vol}{Related Volcano}
  \item{Location Name}{The location of the earthquake}
  \item{Latitude}{The latitude of the earthquake}
  \item{Longitude}{The longitude of the earthquake}
  \item{Focal Depth (km)}{The focal depth of the earthquake in kilometers}
  \item{Mag}{The magnitude of the earthquake}
  \item{MMI Int}{The modified Mercalli Intensity scale}
  \item{Deaths}{How many people died}
  \item{Death Description}{Scale from 0-4 of how many people died}
  \item{Missing}{How many people went missing}
  \item{Missing Description}{Scale from 0-4 of how many people went missing}
  \item{Injuries}{How many people were injured}
  \item{Injuries Description}{Scale from 0-4 of how many people were injured}
  \item{Damage ($Mil)}{Amount of damage in units of million dollars}
  \item{Damage Description}{Scale from 0-4 of how much damage was caused}
  \item{Houses Destroyed}{The amount of houses destroyed in the earthquake}
  \item{Houses Destroyed Description}{Scale from 0-4 of how many houses were destroyed}
  \item{Houses Damaged}{How many houses were damaged}
  \item{Houses Damaged Description}{Scale from 0-4 of how many houses were damaged}
  \item{Total Deaths}{Total amount of deaths in the earthquake}
  \item{Total Death Description}{Scale from 0-4 of how many total deaths there were}
  \item{Total Missing}{Total amount of people who went missing}
  \item{Total Missing Description}{Scale from 0-4 of how many people went missing}
  \item{Total Injuries}{Total amount of people injured}
  \item{Total Injuries Description}{Scale from 0-4 of how many people were injured in total}
  \item{Total Damage ($Mil)}{Total amount of damage in units of million dollars}
  \item{Total Damage Description}{Scale from 0-4 of how much total damage was caused}
  \item{Total Houses Destroyed}{Total amount of houses destroyed}
  \item{Total Houses Destroyed Description}{Scale from 0-4 of how many total houses were destroyed}
  \item{Total Houses Damaged}{Total amount of houses that were damaged}
  \item{Total Houses Damaged Description}{Scale from 0-4 of how many total houses were damaged}
}
}
\source{
DOI:10.7289/V5TD9V7K \url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}
}
\usage{
EarthquakeData
}
\description{
EarthquakeData
}
\keyword{datasets}
