% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Load and cleanup NAOO earthquakes information}
\usage{
eq_clean_data(fileName = system.file("extdata/signif.txt", package =
  "earthquake"))
}
\arguments{
\item{fileName}{Location of file with earthquakes information}
}
\value{
A dataframe with clean information about the earthquakes
}
\description{
Load and cleanup NAOO earthquakes information
}
\examples{
\dontrun{
data <- eq_clean_data()
}

}
