% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates an HTML formatted label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame containing cleaned NOAA earthquake data}
}
\value{
A character vector HTML formatted with earthquakes information
}
\description{
This function creates an HTML formatted label that can be used
for the \code{leaflet} map pop up information. It includes location
name, magnitude and casualties from NOAA earthquake data.
}
\details{
The input \code{data.frame} needs to include columns LOCATION_NAME,
EQ_PRIMARY and TOTAL_DEATHS with the earthquake location, magintude and
total casualties respectively.
}
\examples{
\dontrun{
data \%>\% dplyr::filter(COUNTRY == "SPAIN") \%>\%
  dplyr::mutate(POPUP = eq_create_label(.)) \%>\%
  eq_map(annot_col = "POPUP")
}
}
