% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jggData.R
\name{eq_get_data}
\alias{eq_get_data}
\title{eq_get_data}
\usage{
eq_get_data(filename = NULL)
}
\arguments{
\item{filename}{Path to earthquakes file if already downloaded
when missing or NULL the file will be downloaded from NOAA}
}
\value{
The dataframe generated
        Error if file doesn't exist
}
\description{
load the file
}
