% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jggWrappers.R
\name{eq_load_data}
\alias{eq_load_data}
\title{eq_load_data}
\usage{
eq_load_data(filename = NULL)
}
\arguments{
\item{filename}{Path to earthquakes file if already downloaded
when missing the file will be downloaded from NOAA}
}
\value{
The dataframe generated with columns described here:
        \url{https://www.ngdc.noaa.gov/nndc/struts/results?&t=101650&s=225&d=225}
        Error if file doesn't exist
}
\description{
dowload and clean the earthquakes file from NOAA
             This is a wrapper for eq_get_data and eq_clean_data
}
