% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location information}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{dataframe to clean-up with column LOCATION_NAME}
}
\value{
A dataframe with modified LOCATION_NAME column
}
\description{
Remove country and colomns from location information. Also convert names to title case
}
\examples{
\dontrun{
data <- readr::read_delim("./signif.txt", delim = "\\t") \%>\%
  eq_location_clean()
}
}
