% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Leaflet map of earthquakes}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A data frame containing cleaned NOAA earthquake data}

\item{annot_col}{A character. The name of the column in the data that should
be used for the pop up information. HTML code can be included in this column
for formatted text}
}
\value{
A leaflet map with earthquakes and annotations.
}
\description{
This function creates a \code{leaflet} map of selected earthquakes.
}
\examples{
\dontrun{
data \%>\% dplyr::filter(COUNTRY == "SPAIN") \%>\%
  eq_map(annot_col = "DATE")
}
}
