% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_data_cleaning.R
\name{eq_raw_cleaner}
\alias{eq_raw_cleaner}
\title{A function to clean raw NOAA data}
\usage{
eq_raw_cleaner(data)
}
\arguments{
\item{data}{A tibble of raw NOAA data}
}
\value{
A data.frame of clean NOAA data
}
\description{
This function takes raw NOAA data, downloaded in tab-delimited format
and read into R using \code{readr::read_delim}.
It removes the unnecessary Search Parameters column, and  filters out
rows where Latitude, Longitude, or Year are missing (NA).
It creates a new Date column and a new Country column which are useful
for visualization purposes. It also cleans up the existing Location
Name column and ensures that the Latitude and Longitude columns are
numeric.
}
