% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{eq_read_data}
\alias{eq_read_data}
\title{Read in the NOAA earthquake raw data file}
\usage{
eq_read_data(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file}
}
\value{
This function returns a \code{tbl_df} object.
}
\description{
This function reads in a tab-delimited NOAA earthquake file.
}
\note{
An error message is printed if the file is not found.
}
\examples{
my_raw_data <- eq_read_data(filename="NOAA_earthquakes.txt")

}
