% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataframe.R
\name{eq_read_raw_data}
\alias{eq_read_raw_data}
\title{Imports raw data from file}
\usage{
eq_read_raw_data(file)
}
\value{
list containing imported raw data
}
\description{
Imports raw data from file
}
\examples{
\dontrun{eq_read_raw_data()}
}
