% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.r
\name{eq_timeline}
\alias{eq_timeline}
\title{Create a timeline plot of earthquakes quickly}
\usage{
eq_timeline(dataset, xmin = NA, xmax = NA, label = FALSE)
}
\arguments{
\item{dataset}{A data frame of earthquake data, as returned by
eq_clean_data.}

\item{xmin}{A date object giving a lower bound for the dates represented 
in the plot. If xmin is set, the dataset is subsetted to 
exclude dates before xmin.}

\item{xmax}{A date object giving an upper bound for the dates 
represented in the plot. If xmax is set, the dataset is 
subsetted to exclude dates after xmax.}

\item{label}{A Boolean indicating whether labels showing the local
location of the earthquake should be included in the plot.}
}
\value{
This function creates an an earthquake timeline plot where the
        fill color represents the total number of deaths and the point
        size represents the magnitude.
}
\description{
This function facilitates creating an earthquake timeline plot with
or without location labels using a custom theme.
}
\examples{
eq_data <- eq_clean_data("NOAA_earthquakes.txt")
eq_data <- subset(eq_data, Country == "Mexico" & YEAR >= 2000)

eq_timeline(dataset=eq_data, label=TRUE)

}
