% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.r
\name{eqTheme}
\alias{eqTheme}
\title{Create a theme for earthquake timeline plots.}
\usage{
eqTheme()
}
\value{
This function changes the theme of a ggplot.
}
\description{
This function creates a ggplot theme for earthquake timeline plots.
It is intended to be used with geom_timeline and geom_timeline_label.
}
