% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomTimeLine.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
This is the wrapper to the layer function for the GeomTimeline

The wrapper is just to plot the object. See \code{ggplot2} package documenation
for more specifics on geom_* and layer_* functions.
}
