% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\name{geom_timelinelabel}
\alias{geom_timelinelabel}
\title{geom_timelinelabel}
\usage{
geom_timelinelabel(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping aesthetics to be used for your time line. Includes x, y, label, color, and alpha parameters.}

\item{data}{the data to be used to add your labels. Defaults to the ggplot data if not specified.}

\item{stat}{A stat to be used to transform the given data.}

\item{position}{The positioning of the time line labels on the graph.}

\item{na.rm}{Removes NA's if TRUE.}

\item{show.legend}{Can show a legend for the labels}

\item{inherit.aes}{Asks if you would like to inherit parameters.}

\item{...}{any additional arguments.}
}
\value{
A gTree object with segment and text Grob children added to your graph
}
\description{
This function adds labels to your time line.
}
\note{
In order to label only certain points by an attribute, use n_max() on the data given in the data parameter.
}
\examples{
\dontrun{
ggplot(data = EarthquakeData[1000:1020,]) +
geom_timeline(aes(x = Date, y = 0, color = `Total Deaths`, alpha = Mag)) +
geom_timelinelabel(aes(x = Date, y = 1, label = `Location Name`)) +
ylim(c(-1, 2)) +
theme_minimal()
}
\dontrun{
Earthquake3000 <- arrange(filter(EarthquakeData[3000:3150,], Country \%in\% c("Italy", "Indonesia", "China")), desc(Mag))

ggplot(data = Earthquake3000) +
  geom_timeline(aes(x = Date, y = factor(Country, levels = unique(Country)), color = `Total Deaths`, size = Mag), alpha = 0.5) +
  geom_timelinelabel(data = n_max(Earthquake3000, 11, 5),
                     aes(x = Date, y = factor(Country, levels = unique(Country)), label = `Location Name`)) +
  theme_minimal()}



}
