% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTimeline_Label.R
\docType{data}
\name{GeomTimeLineLabel}
\alias{GeomTimeLineLabel}
\title{GeomTimeLineLabel}
\format{An object of class \code{GeomTimeLineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeLineLabel
}
\description{
This is the GeomTimeLineLabel. It creates the labels for the top
n earthquakes
}
\keyword{datasets}
