% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\alias{GeomTimeline}
\alias{GeomTimelineLabel}
\title{ggproto object for `geom_timeline`}
\description{
This creates the underlying GEOM proto for the timeline geom.

This creates the underlying GEOM proto for the timelinelabel geom.
}
\keyword{datasets}
