% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{IsPosixct}
\alias{IsPosixct}
\title{Test if a charachter string represents a date}
\usage{
IsPosixct(string)
}
\arguments{
\item{string}{Character string with data to test if it is a Posixct date}
}
\value{
A boolean with TRUE if string contains a date and FALSE otherwise
}
\description{
Test if a charachter string represents a date
}
\examples{
\dontrun{
# Returns TRUE
IsPosixct("2010-01-01")

#Returns FALSE
IsPosixct("2010j01-xx")
}
}
