% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jggWrappers.R
\name{jgg_CheckDataType}
\alias{jgg_CheckDataType}
\title{jgg_CheckDataType}
\arguments{
\item{data}{The object which to want check de data type}
}
\description{
Internal function to validate input parameters
}
