% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytheme.R
\docType{data}
\name{mytheme}
\alias{mytheme}
\title{############# earthquake package ##############
This is a package that contains functions to clean and
visualize NOAA earthquake data of countries
###############################################}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 60.}
\usage{
mytheme
}
\description{
4.New theme for the timeline plot
}
\details{
This is a new theme to use with timeline plot
}
\examples{
\dontrun{
p<-ggplot(data2,aes(x=DATE,y=COUNTRY,size=EQ_PRIMARY,fill=TOTAL_DEATHS))+geom_timeline()
p+mytheme+labs(fill="# deaths",size="Richter scale")
}

}
\keyword{datasets}
