% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\name{n_max}
\alias{n_max}
\title{n_max}
\usage{
n_max(df, column, n, desc = TRUE)
}
\arguments{
\item{df}{The data frame you would like to filter}

\item{column}{The number of the column you want to arrange the order of your data by.}

\item{n}{The number of labels you want. The labels will be retrieved from the top of the data you rearranged.}

\item{desc}{TRUE if you want to arrange the data in descending order to label points with the highest numbers.
FALSE if you want to label the points with the lowest numbers.}
}
\value{
The rows of the data frame from which the labels will be retrieved from.
}
\description{
This function filters data for desired labels.
}
\examples{
\dontrun{n_max(EarthquakeData[3000:3150,], 11, 7)}
\dontrun{ggplot(EarthquakeData[6100:6150,]) +
geom_timeline(aes(x = Date, y = factor(Country, levels = unique(Country)), color = Mag)) +
  geom_timelinelabel(data = n_max(EarthquakeData[6100:6150,], 11, 7),
                     aes(x = Date, y = factor(Country, levels = unique(Country)), label = `Location Name`)) +
  ylab("Country") +
  theme_minimal()}

}
