% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\name{read_earthquake_data}
\alias{read_earthquake_data}
\title{read_earthquake_data}
\usage{
read_earthquake_data(file)
}
\arguments{
\item{file}{The file you would like to read in for use.}
}
\value{
Returns the data from your tsv file
}
\description{
Reads in the data from your file with earthquake data. Just calls in read_tsv from readr package, but
specified to earthquake data. Could really be used with any tsv file.
}
\note{
The data required for this package can be obtained at \url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search.}
Click the search button with no filters and the resulting data will contain all the information in their database.
You can then download the tsv file from the download button in the top left corner of the resulting data table.
Citation: DOI:10.7289/V5TD9V7K
}
\examples{
\dontrun{read_earthquake_data("NOAASignificantEarthquakes.tsv")}

}
