% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\name{replace_NA}
\alias{replace_NA}
\title{replace_NA}
\usage{
replace_NA(df, value)
}
\arguments{
\item{df}{The dataframe column to be modified using the df$column syntax}

\item{value}{The value used to replace the NA's in the given column}
}
\value{
The original dataframe with the updated column is returned.
}
\description{
Replaces the NA's in a column of a data frame with a specified value
}
\examples{
\dontrun{replace_NA(EarthquakeData$Mo, 01)}
\dontrun{replace_NA(EarthquakeData$Hr, 00)}

}
