% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataframe.R
\name{setup_data}
\alias{setup_data}
\title{Top-level function to setup the data for analysis}
\usage{
setup_data()
}
\value{
DataFrame with re-formatted NOAA data
}
\description{
Top-level function to setup the data for analysis
}
\examples{
\dontrun{setup_data()}
}
