% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{stat_timeline_label}
\alias{stat_timeline_label}
\title{Create timeline label stat function}
\usage{
stat_timeline_label(
  mapping = NULL,
  data = NULL,
  geom = "timelineLabel",
  position = "identity",
  show.legend = TRUE,
  inherit.aes = TRUE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Defaults to NULL}

\item{data}{Defaults to NULL, so the data is inherited from the plot
data as specified in the \code{ggplot2::ggplot()} call}

\item{geom}{Defaults to "timelineLabel"}

\item{position}{Defaults to "identity"}

\item{show.legend}{Defaults to TRUE}

\item{inherit.aes}{Defaults to TRUE}

\item{na.rm}{Defaults to TRUE}

\item{...}{Arguments to be passed to methods}
}
\value{
a layer which renders the data to make it suitable for the
timeline label geom to plot
}
\description{
This creates the \code{stat_timeline_label()} function.
It uses the ggproto object StatTimelineLabel.
The \code{stat_timeline_label} function takes data of a certain
format, processes it, and returns it in the correct format for the
timeline label geom to plot.
}
\section{Aesthetics}{

\code{stat_timeline_label()} understands the following aesthetics:
\itemize{
\item{x}{REQUIRED. A column of dates. Column type must be Date}
\item{xmin}{REQUIRED. Minimum/starting date. Must be a Date object}
\item{xmax}{REQUIRED. Maxmimum/ending date. Must be a Date object}
\item{txt}{REQUIRED. Any type of column to be displayed as text}
\item{max_by}{REQUIRED. A numeric/Date column to select maximum
observations by}
\item{n_max}{optional, if not supplied will be auto calculated.
A number specifying how many values should be selected from the
max_by column (arranged in descending order) }
}
}

\examples{
  \dontrun{ggplot2::ggplot(data=
                                eq_raw_cleaner(tail(eq,100)),
                           ggplot2::aes(x=Date,
                                        xmin=as.Date("2022-01-14"),
                                        xmax=as.Date("2022-05-26"),
                                        max_by=Mag,
                                        txt=Country,
                                        n_max=3)) +
                           stat_timeline_label()}

}
