% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_timeline.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{Timeline Stat}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatTimeline
}
\description{
Timeline Stat
}
\keyword{datasets}
