% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{timeline_draw}
\alias{timeline_draw}
\title{Draw function for \code{geom_timeline}}
\usage{
timeline_draw(coords)
}
\arguments{
\item{coords}{A data.frame passed from \code{stat_timeline} to
\code{geom_timeline} and transformed for plotting.}
}
\value{
a gTree object containing the grobs to be drawn
}
\description{
This function takes the data.frame returned by \code{stat_timeline}
which has already been transformed for plotting.
It then uses the columns of the data.frame to create three types of
grob: a line that lies across the x-axis (the timeline), points that
represent individual earthquake observations, and segments that go
through each point to clarify its position on the y-axis (this is
helpful when a continuous y-aesthetic has been specified).
The points are colored, filled, and sized according to the colour,
fill, alpha, and size aesthetic specifications. If any/all of these
aesthetics are not specified by the user, default aesthetics are
provided in the \code{geom_timeline} code.
The function then creates a gTree out of these grobs which it returns
to be plotted.
}
\examples{
  \dontrun{timeline_draw(data.frame(x=c(7,12),
                                    xmin=c(6,6),
                                    xmax=c(15,15),
                                    y=c(3,6),
                                    size=c(5,5),
                                    colour=c(1,1),
                                    fill=c(1,1),
                                    alpha=c(1,1))}

}
