% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{timeline_label_draw}
\alias{timeline_label_draw}
\title{Draw function for \code{geom_timeline_label}}
\usage{
timeline_label_draw(coords)
}
\arguments{
\item{coords}{A data.frame passed from \code{stat_timeline} to
\code{geom_timeline} and transformed for plotting.}
}
\value{
a gTree object containing the grobs to be drawn
}
\description{
This function takes the data.frame returned by
\code{stat_timeline_label} which has already been transformed for
plotting.
It then uses the columns of the data.frame to create up to five grob
types: a segment that lies across the x-axis (timeline), points that
represent individual earthquake observations, a segment that goes
through the points, text displaying the information in whichever
column the user has supplied to the txt argument, and a segment
that joins the text to the point to which it corresponds.
Text grobs and corresponding segments are only produced for a few of
the largest values in the column supplied to the max_by aesthetic.
The exact number of text grobs to be produced is determined by the
optional aesthetic n_max.
The points are colored, filled, and sized according to the colour,
fill, alpha, and size aesthetic specifications. If any/all of these
aesthetics are not specified by the user, default aesthetics are
provided in the \code{geom_timeline} code.
The function then creates a gTree out of these grobs which it returns
to be plotted.
}
