% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_data_cleaning.R
\name{year_pad}
\alias{year_pad}
\title{A function to deal with 2- or 3- digit year dates}
\usage{
year_pad(y, m = 1, d = 1)
}
\arguments{
\item{y}{A year number as a character string}

\item{m}{A month number (1-12) as a character string. Defaults to 1
(January)}

\item{d}{A day number as a character string. Defaults to 1 (the 1st).}
}
\value{
An AD/CE date of class Date in y-m-d format
}
\description{
This function handles 2- or 3- digit year dates.
It takes as input a year, month, and day. If no month and day are
inputted, the month defaults to the first month (January) and the day
defaults to the first (1st).
The function turns a 2- or 3- digit year into a 4-digit year by
padding the left side with zeros. It does not change years which are
already 4 digits long.
}
\examples{
\dontrun{year_pad(y=28)}
\dontrun{year_pad(y=123, m=5,d=22)}

}
