% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning of earthquake data}
\usage{
eq_clean_data(text)
}
\arguments{
\item{text}{name of data set as string}
}
\value{
This function returns the earthquake data as a list
}
\description{
This is a function that reads the data from a the csv file containing the earthquake data. It create a new date column, cleans the names where the earthquake
occured and converts longitude and latitude values in to numeric types.
}
\examples{
\dontrun{
data<-eq_clean_data("earthquakes_data.txt.zip")
}

}
