% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean up a raw NOAA earthquake data file}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file}
}
\value{
This function returns a \code{tbl_df} object.
}
\description{
This function cleans up a NOAA earthquake file.
}
\note{
An error message is printed if the file is not found.
}
\examples{
my_clean_data <- eq_clean_data(filename="signif.txt")

}
