% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{print("eq_clean_data")
This function applies some cleaning processes to a dataset downloaded from
the \href{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}{NOAA Significant Earthquake Database}.
Specifically, it converts the Latitude and Longitude columns to numeric class
and combines year, month and day into a single variable in the Date class.}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{the raw data frame to clean}
}
\value{
This function returns a dataset where Longitude and Latitude are
numeric variables and there is a new Date column containing the date. The old
year, month and day columns are dropped.
}
\description{
print("eq_clean_data")
This function applies some cleaning processes to a dataset downloaded from
the \href{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}{NOAA Significant Earthquake Database}.
Specifically, it converts the Latitude and Longitude columns to numeric class
and combines year, month and day into a single variable in the Date class.
}
\examples{
\dontrun{
cleandata <- eq_clean_data(rawdata)
}

}
