% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jggData.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
The dataframe/datatable cleaned
}
\description{
Clean and prepare the dataframe for future uses
             Cleaning involves:
             1.- remove unused columns
             2.- Convert latitude and longitude to numeric
             3.- create the DATE field
             4.- Adjust LOCATION_NAME with eq_location_clean
}
\examples{
\dontrun{
dt = eq_clean_data(dataframe)
}
}
