% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{earthquake package}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{a dataframe or tibble to be cleaned}
}
\value{
a cleaned dataframe or tibble
}
\description{
2.Clean earthquake data with \code{eq_clean_data} function
}
\details{
This function is used to clean variables to be used in
the plots.It will generate a DATE variable of class Date
based on YEAR, MONTH and DAY in the data set. It will
convert LATITUDE, LONGITUDE, EQ_PRIMARY, TOTAL_DEATHS to
numeric and clean LOCATION_NAME using eq_location_clean().
}
\examples{
\dontrun{
data<-eq_clean_data(data)
}

}
