% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{The data frame to be cleaned. Specified for data from the NOAA Significant Earthquake Database,
but can be used for any dataframe with the correct column names}
}
\value{
Returns the given data frame in a cleaner version with the year, month, and date in a single column
as a Date object. Also converts Latitude and Longitude columns into the numeric class.
}
\description{
This function cleans the data from the NOAA Significant Earthquake Database.
}
\note{
NA's in the Mo and Dy columns are replaced by 01. NA's in the Hr, Mn, and Sec columns are replaced with 0.
}
\examples{
\dontrun{eq_clean_data(EarthquakeData)}

}
