% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data
loads earthquake data.}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{earthquakes}{the file that contains the earthquake data}
}
\value{
Returns earthquake data with the following cleaning -a- added 01 for
empty months and days and years -b- removed dates in bc -c- created a DATE
column with date format -d- turned long and lat to numeric -d- use functions
to clean Location Names and Country Names
}
\description{
eq_clean_data
loads earthquake data.
}
\examples{
eq_clean_data("earthquakes-2020-10-19_15-21-05_+0300.tsv")
}
