% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataframe.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Creates a DataFrame with filled in DATE value and properly formatted LONGITUDE, LATITUDE and}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{list containing raw NOAA data}
}
\value{
DataFrame containing formatted NOAA data
}
\description{
Creates a DataFrame with filled in DATE value and properly formatted LONGITUDE, LATITUDE and
}
\examples{
\dontrun{eq_clean_data(raw_data)}
}
