% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanup.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleanup Earthquake Data}
\usage{
eq_clean_data(raw_data, YEAR = "YEAR", MONTH = "MONTH", DAY = "DAY",
  LONGITUDE = "LONGITUDE", LATITUDE = "LATITUDE",
  MAGNITUDE = "EQ_PRIMARY", DEATH = "DEATHS")
}
\arguments{
\item{raw_data}{The earthquake dataframe that needs to be cleaned.}

\item{YEAR}{The name of the column that holds the year of the earthquake as a numeric/integer.}

\item{MONTH}{The name of the column that holds the month of the earthquake as a numeric/integer.}

\item{DAY}{The name of the column that holds the day of the month of earthquake as a numeric/integer.}

\item{LONGITUDE}{The longitude of the earthquake - numeric.}

\item{LATITUDE}{The latitude of the earthquake - numeric.}

\item{MAGNITUDE}{The latitude of the earthquake - numeric or character.}

\item{DEATH}{The latitude of the earthquake - numeric or character.}
}
\value{
A cleaned up dataframe with the new date column.
}
\description{
This functions cleans up data from the NOAA significant earthquake database.
The cleanup includes building a single date field
and removing BCE dates since POSIXlt dates do not work well with pre 1 CE dates.
}
\examples{
\donttest{
read_delim(file = 'signif.txt', delim = '\\t') \%>\% earthquake::eq_clean_data()
}

}
