% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the raw NOAA earquake data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{an input dataframe

return a clean dataframe}
}
\description{
This function performs two task: merge YEAR, MONTH, DAY into one DATE column
and convert LONGITUDE AND LATITUDE into numeric.
}
\examples{
fname <- system.file("extdata", "signif.txt.tsv", package = "earthquake")
eq_dat_raw <- as.data.frame(readr::read_tsv(fname)) 
eq_dat_clean <- eq_clean_data(eq_dat_raw)
}
