% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_country_clean}
\alias{eq_country_clean}
\title{eq_country_clean
cleans location column only leaving the country}
\usage{
eq_country_clean(loc_col)
}
\arguments{
\item{loc_col}{the column that holds the location data}
}
\value{
returns a data file where location is -a- trimmed of ws -b- converted
to title case -c- name of country removed
}
\description{
eq_country_clean
cleans location column only leaving the country
}
\examples{
\dontrun{
eq_country_clean("Canada:Toronto")
}
}
