% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.r
\name{eq_create_label}
\alias{eq_create_label}
\title{Create information for pop-up windows on an earthquake map}
\usage{
eq_create_label(dataset)
}
\arguments{
\item{dataset}{A dataset generated by eq_clean_data().}
}
\value{
A string containing HTML code to display the location, 
        information is missing, the HTML object for that value is 
        omitted.
}
\description{
This function creates HTML content for the pop-up windows on an 
earthquake map. The input to this function should be a data set 
generated by the function eq_clean_data(). The window will show 
the local location of the earthquake (LocalLocation), the magnitude 
(EQ_PRIMARY), and the total death count (TOTAL_DEATHS).
}
\examples{
eq_data <- eq_clean_data("NOAA_earthquakes.txt") \%>\% 
  dplyr::filter(Country == "Mexico" & YEAR >= 2000)

eq_data \%>\% 
  dplyr::filter(COUNTRY == "MEXICO" & YEAR >= 2000) \%>\% 
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\% 
  eq_map(annot_col = "popup_text")

}
