% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{eq_create_label}
\alias{eq_create_label}
\title{print("eq_create_label")
This function creates an HTML label as a new column in an earthquake daframe
where Location, Magnitude and Total Deaths are introduced and that can be
used for Leafle map annotation}
\usage{
eq_create_label(data = rlang::.data)
}
\arguments{
\item{data}{the data frame to add the HTML label to}
}
\value{
This function returns the input dataset with an additional column
called popup_text which contains an HTML string that can be used to
annotate in Leaflet maps  the Location, Magnitude and Total Deaths of
earthquakes.
}
\description{
print("eq_create_label")
This function creates an HTML label as a new column in an earthquake daframe
where Location, Magnitude and Total Deaths are introduced and that can be
used for Leafle map annotation
}
\examples{
\dontrun{
annotateddata <- eq_create_label(data)
}

}
