% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_mapping.R
\name{eq_create_label}
\alias{eq_create_label}
\title{A function to create data labels}
\usage{
eq_create_label(mapdata)
}
\arguments{
\item{mapdata}{A data.frame of data to map}
}
\value{
A column of information written in HTML format which will
appear as a popup on a map.
}
\description{
This function takes a data.frame of data that is to be mapped with
\code{eq_map()}. It creates a column containing popup information
written in HTML format. This information includes the location name,
the earthquake magnitude, and total deaths resulting.
If any of this information is missing for a certain observation, it
will not be displayed in the popup.
}
\examples{
\dontrun{data<-read.delim(system.file("extdata","noaa_earthquakes.tsv", package="earthquake")),header=TRUE
         data<-eq_raw_cleaner(data)
         data<-dplyr::mutate(data, popup_text=eq_create_label(data)}

}
