% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{############# earthquake package ##############
This is a package that contains functions to clean and
visualize NOAA earthquake data of countries
###############################################}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a data frame or tibble}
}
\value{
a data frame with a new column popup_text
}
\description{
7.Create label of earthquakes for annotation
}
\details{
\code{eq_create_label} creates labels that contain
location, magnitude and deaths of earthquakes.
}
\examples{
\dontrun{
eq_create_label(data2) \%>\% eq_map("popup_text")
}

}
