% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The data frame to retrieve data from}
}
\value{
A vector of HTML labels to add to your data frame. Labels include Location, Magnitude, and Total Deaths.
}
\description{
Creates a label in HTML format for each point observation of your data.
}
\note{
Data must contain columns `Location Name`, Mag, and `Total Deaths` for this function to work.
}
\examples{
\dontrun{
EarthquakeData[EarthquakeData$Country == "Usa", colnames(EarthquakeData)] \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = .$popup_text)}
\dontrun{
EarthquakeData \%>\%
  filter(Country == "Italy" & year(Date) >= 2000) \%>\%
  mutate(popup_label = eq_create_label(.)) \%>\%
  eq_map(annot_col = .$popup_label, radius = 7, weight = 1)}

}
