% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label
produces a column in the earthquake data that contains the information
necessary for a popu-up - location, magnitude, total deaths}
\usage{
eq_create_label(df)
}
\arguments{
\item{data}{The earthquake dataframe}
}
\value{
string with labels for the popup
}
\description{
eq_create_label
produces a column in the earthquake data that contains the information
necessary for a popu-up - location, magnitude, total deaths
}
\examples{
library(dplyr)
eq = eq_clean_data("earthquakes-2020-10-19_15-21-05_+0300.tsv")
eq = dplyr::mutate(eq, popup_text = eq_create_label(eq))
eq_map(eq, annot_col = "popup_text")
}
