% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_earthquake.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create earthquak label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A column of earthquak annotation; locaion, magnitude and deaths. Ignores the fiels when NA.
}
\description{
Create earthquak label
}
\examples{
\dontrun{
# load and clean data
file_path <- system.file("extdata", "signif.tsv", package = "MSDR")
signif <- read_tsv(file_path) \%>\%
  eq_clean_date \%>\%
  eq_clean_location

# make graph
signif \%>\%
filter(COUNTRY == 'MEXICO' & year(date) >= 2000) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map()
}
}
