% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_plots.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create fancy annotations for the leaflet plots}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A dataframe that contains the columns: LOCATION_NAME, EQ_PRIMARY, and DEATHS}
}
\value{
A character column with html-tagged annotations
}
\description{
This function reads in a dataframe with LOCATION_NAME, EQ_PRIMARY, and DEATHS as columns.
It then creates and html tagged annotation for use in the leaflet plots.
}
\examples{
\donttest{
data(earthquake_raw)
earthquake_raw \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")}

}
