% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletmap.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create popup information}
\usage{
eq_create_label(dat)
}
\arguments{
\item{dat}{an imput dataframe}
}
\value{
an interactive map with three types of information on popup if existing.
}
\description{
This function generates html formated earthquake information by conbination 
three types of data, including location, magnitude and number of total deaths.
}
