% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{print("eq_location_clean")
This function applies some cleaning processes to the Location Name column in
a dataset downloaded from the \href{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}{NOAA Significant Earthquake Database}.
Specifically, it drops any location beyond that or those of the first
occurring country, it drops the preceding name of country and region, and
converts the resulting location name to title case.}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{the raw data frame to clean}
}
\value{
This function returns a dataset where the Location Name column has
been cleaned, by removing location in more that one country, removing the
country name and converting the name to title case. The country name, where
available is also converted to title case and assigned to a separate column.
}
\description{
print("eq_location_clean")
This function applies some cleaning processes to the Location Name column in
a dataset downloaded from the \href{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}{NOAA Significant Earthquake Database}.
Specifically, it drops any location beyond that or those of the first
occurring country, it drops the preceding name of country and region, and
converts the resulting location name to title case.
}
\examples{
\dontrun{
cleandata <- eq_location_clean(data)
}

}
