% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jggData.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
The dataframe/datatable modified
}
\description{
Create LOATION_NAME based on country
}
\examples{
\dontrun{
dt = eq_location_clean(dataframe)
}
}
