% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function to clean Location Name column}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A tibble passed to the function through the pipe chain
within the \code{eq_raw_cleaner} function.}
}
\value{
A tibble with the original Location Name column replaced by
a cleaned Location Name column
}
\description{
This function cleans the Location Name column of the NOAA data by
stripping away the country name which is placed before the colon :.
It keeps the location name that follows the colon :.
It also converts the column type from character to factor to
facilitate data visualization later on.
}
\examples{
  \dontrun{eq_location_clean(eq_raw_cleaner(tail(eq,100)))}

}
