% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{############# earthquake package ##############
This is a package that contains functions to clean and 
visualize NOAA earthquake data of countries
###############################################}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{a string to be cleaned, with country name 
and location name separated by colon (:).}
}
\value{
a string that is cleaned, if location name only
contains country name, then country name will be returned
}
\description{
1.Clean location name with \code{eq_location_clean} function
}
\details{
This function is used to clean LOCATION_NAME variable by
stripping out country names and keep the location name, 
which will be used in plots. The location name will be 
convert to title case.
}
\examples{
\dontrun{
eq_location_clean("ECUADOR:  RIOBAMBA, QUITO, CUZCO")
eq_location_clean("TURKEY")
eq_location_clean("China: Jiangsu: Nanjing")
}

}
