% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df, j)
}
\arguments{
\item{df}{The data frame to be cleaned.}

\item{j}{The column number or name for the "location name" column in your data frame.}
}
\value{
Returns the data frame with the location name column updated to remove countries and adding a new column for countries.
}
\description{
This function cleans the Location Name column of a given data set,
removing the country name along with the colons and white spaces around the city/area name.
A separate column for countries is added.
}
\examples{
\dontrun{eq_location_clean(EarthquakeData, 7)}

}
