% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean
cleans location column and removes country data}
\usage{
eq_location_clean(loc_col)
}
\arguments{
\item{loc_col}{the column that holds the location data}
}
\value{
returns a data file where location is -a- trimmed of ws -b- converted
to title case -c- name of country removed
}
\description{
eq_location_clean
cleans location column and removes country data
}
\examples{
\dontrun{
eq_location_clean("Canada:Toronto")
}
}
