% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataframe.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the LOCATION_NAME of a DataFrame containing NOAA data}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{DataFrame containing partially-cleaned NOAA data}
}
\value{
DataFrame with reformatted LOCATION_NAME column
}
\description{
Cleans the LOCATION_NAME of a DataFrame containing NOAA data
}
\examples{
\dontrun{eq_location_clean(df)}
}
