% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanup.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Location name cleanup}
\usage{
eq_location_clean(my_data, LOCATION_NAME = "LOCATION_NAME")
}
\arguments{
\item{my_data}{data.frame that holds your NOAA earthquake data. Can be raw or cleaned up by earthquake::eq_clean_data}

\item{LOCATION_NAME}{A character with the name of the column that holds the raw NOAA location name.}
}
\description{
This function drops the country-name and colon from the location name.
Furthermore, it converts the text to title case. WOW, it does soooo much!!!
}
\examples{
\donttest{
data(earthquake_raw)
earthquake_raw \%>\% earthquake::eq_location_clean()
}

}
