% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Help function for cleanind raw data}
\usage{
eq_location_clean(loc)
}
\arguments{
\item{loc}{LOCATION_NAME vector

return cleaned LOCATION_NAME}
}
\description{
This function speciallly extracts location information from
LOCATION_NAME column
}
\examples{
fname <- system.file("extdata", "signif.txt.tsv", package = "earthquake")
eq_dat_raw <- as.data.frame(readr::read_tsv(fname)) 
locations <- eq_location_clean(eq_dat_raw$LOCATION_NAME)
locations

}
