% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Mapping of earthquake data}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{a cleaned data file}

\item{annot_col}{name of column used for annotations}
}
\value{
map of the earthquakes including annotations
}
\description{
This function plots the earthquake and adds some annotations to it
}
\examples{
\dontrun{
readr::read_delim("earthquakes_data.txt.zip", delim = "\\t") \%>\%
 eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(datetime) >= 2000)\%>\%
  eq_map(annot_col = "datetime")
}

}
