% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.r
\name{eq_map}
\alias{eq_map}
\title{Create a leaflet map showing earthquake locations}
\usage{
eq_map(dataset, annot_col)
}
\arguments{
\item{dataset}{A dataset generated by eq_clean_data().}

\item{annot_col}{A string giving the name of a column in dataset. The 
contents of this will be used to fill in the 
pop-up window when an earthquake is selected.}
}
\value{
This function creates a leaflet map with circles 
        representing the epicenters of the earthquakes. The radius of 
        each circle is determined by the field EQ_PRIMARY.
}
\description{
This function creates a leaflet map plot showing the epicenter of 
each earthquake in a dataset.
}
\examples{
eq_data <- eq_clean_data("NOAA_earthquakes.txt") \%>\% 
  dplyr::filter(Country == "Mexico" & YEAR >= 2000)

eq_map(dataset=eq_data, annot_col = "DATE")

}
