% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{eq_map}
\alias{eq_map}
\title{print("eq_map")
This function creates an interactive Leaflet map to visualize earthquake
locations contained in the input dataframe marking them with a circle whose
size is proportional to the earthquake's magnitude. Earthquake locations can
be annotated}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{the data frame to use for map creation}

\item{annot_col}{a string indicating the name of the column used for popup
annotation}
}
\value{
This function creates a Leaflet map to visualize the earthquake
locations contained in the input dataframe
}
\description{
print("eq_map")
This function creates an interactive Leaflet map to visualize earthquake
locations contained in the input dataframe marking them with a circle whose
size is proportional to the earthquake's magnitude. Earthquake locations can
be annotated
}
\examples{
\dontrun{
eq_map(eqdata, annot_col="Date")
eq_map(eq_create_label(eqdata), annot_col="popup_text")
}

}
