% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jggGeom.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(dt, annot_col)
}
\arguments{
\item{dt}{A data table containing NOAA Earthquake data cleaned}

\item{annot_col}{A column found in \code{eq_data} to annotate earthquake marker}
}
\value{
An interactive map displaying earthquate location for a given country with user defined popup.
}
\description{
A function to generate an interactive map showing earthquakes.
The user specifies a column from the data which the earthquake is to be annotated by eg DATE.
}
\examples{
\dontrun{
library(dplyr)
library(lubridate)
dt <- eq_clean_data \%>\% filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000)
eq_map(dt, annot_col="DATE")
}
}
