% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_mapping.R
\name{eq_map}
\alias{eq_map}
\title{A function to map earthquake data}
\usage{
eq_map(mapdata, color = viridis::magma(1), annot_col = mapdata$`Location Name`)
}
\arguments{
\item{mapdata}{Data to be mapped}

\item{color}{The column to be colored by, indicated using the $
operator. Defaults to no column, and every point is colored the
same color in the magma color palette}

\item{annot_col}{The column which contains text for popups, indicated
using the $ operator. Defaults to the Location Name column.}
}
\value{
NULL. Displays a map in the Viewer pane.
}
\description{
A function to map earthquake data
}
\examples{
\dontrun{data<-read.delim(system.file("extdata","noaa_earthquakes.tsv", package="earthquake")),header=TRUE
         data<-eq_raw_cleaner(data)
         eq_map(data)}

}
