% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_mapping.R
\name{eq_map}
\alias{eq_map}
\title{A function to map earthquake data}
\usage{
eq_map(mapdata, color = viridis::magma(1), annot_col = mapdata$`Location Name`)
}
\arguments{
\item{mapdata}{Data to be mapped}

\item{color}{The column to be colored by, indicated using the $
operator. Defaults to no column, and every point is colored the
same color in the magma color palette}

\item{annot_col}{The column which contains text for popups, indicated
using the $ operator. Defaults to the Location Name column.}
}
\value{
NULL. Displays a map in the Viewer pane.
}
\description{
This function creates an interactive Leaflet map of earthquake data.
The locations of points on the map show the earthquake epicenters.
The sizes of points on the map show the earthquake magnitudes.
The points can be colored by a user-specified variable, and a legend is
included to explain the colors.
The user can also specify a column of the earthquake data to be shown
as a popup when each point is clicked.
}
\examples{
\dontrun{eq_map(eq_raw_cleaner(eq))}

}
