% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{earthquake package}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{annot_col}{a character specifies the annotation variable}
}
\value{
an interactive map with earthquakes marked as circles
}
\description{
6.Visualize earthquakes on an interactive map
}
\details{
\code{eq_map} function is used to visualize earthquakes with
annotations on an interactive map.

when click on the earthquake circle, text will pop up
}
\examples{
\dontrun{
eq_map(data2,"DATE")
}

}
