% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_map}
\alias{eq_map}
\title{eq_map
Produces a map of earthquake epicenters and annotates
each point with a popup window}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{The earthquake data to be plotted. See the example for details.}

\item{annot_col}{the column that contains the annotations}
}
\value{
leaflet object with earthquake data
}
\description{
eq_map
Produces a map of earthquake epicenters and annotates
each point with a popup window
}
\examples{
library(dplyr)
eq = eq_clean_data("earthquakes-2020-10-19_15-21-05_+0300.tsv")
eq = dplyr::filter(eq, COUNTRY == "Mexico" & lubridate::year(DATE) >= 2000)
eq_map(eq, annot_col = "COUNTRY")
}
