% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_plots.R
\name{eq_map}
\alias{eq_map}
\title{Create leaflet plot with annotations}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A dataframe that contains the annotation column and numeric columns named LATITUDE and LONGITUDE}

\item{annot_col}{The name of the column that holds the annotations.}
}
\value{
A leaflet plot with annotations.
}
\description{
This function reads in a dataframe and annotation column and outputs a leaflet plot.
}
\examples{
\donttest{
data(earthquake_raw)
earthquake_raw \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")}

}
