% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletmap.R
\name{eq_map}
\alias{eq_map}
\title{leaflet map to show column informtion}
\usage{
eq_map(dat, annot_col)
}
\arguments{
\item{dat}{an imput dataframe}

\item{annot_col}{a string representing colume}
}
\value{
An interactive map with information shown
 on popup
}
\description{
This interactive map can show earquake information on popup
 accoring to given column
}
