% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plotting the earthquake data along time axis}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{if set to 'FALSE' the missing values are removed with a warning.
If set to `TRUE`, the values are removed silently.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{These are other arguments passed to [layer()]. E.g. this can be 'color=red'.}
}
\value{
The geom \code{geom_timeline} used with the \code{ggplot} function, produce a time line plot of selected earthquakes.
}
\description{
This is a ggplot function that displays the data of earthquakes along an x-axis. The magnitude of the earthquake is displayed via the circle radius
of each single data point. The color of each circle is symbolide based on the number of total deaths per earthquake
}
\examples{
\dontrun{
data \%>\%
   mutate(EQ_PRIMARY = round(EQ_PRIMARY)) \%>\%
   filter(datetime>"2010-01-01" & datetime<"2012-01-01")\%>\%
   ggplot(aes(x=datetime, color=DEATHS, size = EQ_PRIMARY, label = LOCATION))+
   geom_timeline(alpha=.8)
}

}
