% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.r
\name{geom_timeline}
\alias{geom_timeline}
\title{Create a timeline plot of earthquakes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{An aesthetic mapping created by ggplot2::aes.}

\item{data}{A data frame containing the data to be plotted.}

\item{stat}{A string indicating the data transformation to be used.}

\item{position}{A string giving the poistion adjustment.}

\item{na.rm}{A logical controlling how missing values are handled. If 
FALSE, missing values are removed and a warning is printed, 
otherwise missing values are removed without a warning.}

\item{show.legend}{A logical indicating whether the layer created by this 
function should have a legend.}

\item{inherit.aes}{A logical. If FALSE, the default aesthetics are 
overridden.}

\item{...}{additional arguments passed to the layer.}
}
\value{
This function adds an earthquake timeline plot to the current
        ggplot.
}
\description{
This function creates a plot showing earthquakes by date and country. 
It uses the GeomTimeline geom. The arguments are similar to those use 
in ggplot2 functions such as geom_point().
}
\examples{
eq_data <- eq_clean_data("NOAA_earthquakes.txt")
eq_data <- subset(eq_data, Country == "Mexico" & YEAR >= 2000)

ggplot2::ggplot(eq_data) +
  geom_timeline(ggplot2::aes(x=DATE, y=Country, fill=TOTAL_DEATHS, 
                 size=EQ_PRIMARY))

}
