% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.r
\name{geom_timeline}
\alias{geom_timeline}
\title{Create a timeline plot of earthquakes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{A vector of dates on which earthquakes occurred.}

\item{y}{A factor vector giving the country in which each earthquake 
occurred (optional).}

\item{shape}{A numeric vector giving the point shape to be used.}

\item{size}{A numeric vector representing the size of each point.}

\item{colour}{A string vector controlling the color of the point borders.}

\item{fill}{A string vector controlling the fill color of the points.}

\item{alpha}{A numeric constant in the interval [0,1] controlling the 
transparency of the wind radii chart. This parameter has a  
default value of 0.4.}

\item{stroke}{A numeric vector controlling the border with of the point.}
}
\value{
This function adds an earthquake timeline plot to the current 
        ggplot.
}
\description{
This function creates a plot showing earthquakes by date and 
 country.
}
\examples{
eq_data <- eq_clean_data("NOAA_earthquakes.txt") \%>\% 
  dplyr::filter(Country == "Mexico" & YEAR >= 2000)

ggplot2::ggplot(eq_data) +
  geom_timeline(aes(x=DATE, y=Country, fill=TOTAL_DEATHS, size=EQ_MAG_MW)) + 
  labs(fill="# deaths", size="Magnitude") + 
  eqTheme()

}
