% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create timeline geom function}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "timeline",
  position = "identity",
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Defaults to NULL}

\item{data}{Defaults to NULL, so the data is inherited from the plot
data as specified in the \code{ggplot2::ggplot()} call}

\item{stat}{Defaults to "timeline", specifying that the statistical
transformation to be used on the data for this layer is the
timeline stat}

\item{position}{Defaults to "identity"}

\item{na.rm}{Defaults to FALSE}

\item{show.legend}{Defaults to TRUE}

\item{inherit.aes}{Defaults to TRUE}
}
\value{
a layer which plots the data as per the GeomTimeline
object definition that can be added to a ggplot object
}
\description{
This creates the \code{geom_timeline()} function.
It uses the ggproto object GeomTimeline.
This function allows the timeline and points (as specified in the
creation of the object GeomTimeline) to be plotted as a ggplot
layer.
}
\section{Aesthetics}{

\code{geom_timeline()} understands the following aesthetics:
\itemize{
\item{x}{REQUIRED. A column of dates. Column type must be Date}
\item{xmin}{REQUIRED. Minimum/starting date. Must be a Date object}
\item{xmax}{REQUIRED. Maxmimum/ending date. Must be a Date object}
\item{y}{optional, defaults to 0.3}
\item{colour}{optional, defaults to black. Column to color points
by}
\item{fill}{optional, defaults to black. Column to fill points by}
\item{size}{optional, defaults to 0pt. Column to size points by}
\item{alpha}{optional, defaults to 1 (no transparency)}
}
}

