% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{geom_timeline}
\alias{geom_timeline}
\alias{GeomTimeline}
\title{earthquake package}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = F,
  show.legend = NA,
  inherit.aes = T,
  ...
)

GeomTimeline
}
\arguments{
\item{mapping}{Set of asthetic mappings created by [aes()]}

\item{data}{data frame}

\item{stat}{stat function=identity}

\item{position}{position=identity}

\item{na.rm}{remove NA=F}

\item{show.legend}{show legend=NA}

\item{inherit.aes}{inherit ase=TRUE}

\item{...}{Other arguments passed on to [layer()].}
}
\description{
3.Timeline plot with \code{geom_timeline} function
}
\details{
This new geom function is used to build a timeline plot
of earthquakes.

asthetics: x is required, specifies date variable to plot on x axis;
y is optional, specifies country variable to plot on y axis;
size is optional, specifies size of the point (pch shape=21),
i.e. magnitude of earthquake;
fill is optional, specifies fill of the point, i.e. deaths.
}
\examples{
\dontrun{
p<-ggplot(data2,aes(x=DATE,y=COUNTRY,size=EQ_PRIMARY,fill=TOTAL_DEATHS))+geom_timeline()
}

}
\keyword{datasets}
