% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline
plots a time line of earthquakes with a point for each earthwquake. x axis is
date and y axis is a factor}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data}{The earthquake data to be plotted. See the example for details.}
}
\value{
ggplot2 graphical object of earthquakes
}
\description{
geom_timeline
plots a time line of earthquakes with a point for each earthwquake. x axis is
date and y axis is a factor
}
\examples{
library(ggplot2)
eq = eq_clean_data("earthquakes-2020-10-19_15-21-05_+0300.tsv")
eq = dplyr::filter(eq, COUNTRY == "Mexico" & lubridate::year(DATE) >= 2000)
ggplot2::ggplot(eq,aes(x=DATE,y = COUNTRY, color = TOTAL_DEATHS, size = TOTAL_DEATHS)) +
  geom_timeline(alpha=.5)
}
