% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.r
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Create a timeline plot of earthquakes with labels showing the 
city or region.}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n_max = NULL, ...)
}
\arguments{
\item{n_max}{A integer constant. If n_max is used, only the n_max 
largest earthquakes, as determined by the value of size, 
are given labels. The default behavior is to label all 
earthquakes.}

\item{x}{A vector of dates on which earthquakes occurred.}

\item{y}{A factor vector giving the country in which each earthquake 
occurred (optional).}

\item{shape}{A numeric vector giving the point shape to be used.}

\item{size}{A numeric vector representing the size of each point.}

\item{colour}{A string vector controlling the color of the point borders.}

\item{fill}{A string vector controlling the fill color of the points.}

\item{alpha}{A numeric constant in the interval [0,1] controlling the 
transparency of the wind radii chart. This parameter has a  
default value of 0.4.}

\item{stroke}{A numeric vector controlling the border with of the point.}

\item{label}{A string vector containing the label that should be used 
for each earthquake; this is usually a city or region name.}
}
\value{
This function an earthquake timeline plot with labels to the 
         current ggplot.
}
\description{
This function creates a plot showing earthquakes by date, country, and 
city or region. It is similar to geom_timeline, but has the option to 
add labels showing the city or region in which the earthquake took 
place.
}
\examples{
eq_data <- data.frame("Date"=as.Date(c("2000-01-01","2010-05-15")),
                      "Country"=c("France","Japan"),
                      "Total_Deaths"=c(0,10),
                      "Magnitude"=c(6.8, 5.6),
                      "City"=c("Paris","Tokyo"))

ggplot2::ggplot(data = eq_data) +
 geom_timeline_label(aes(x=Date, y=Country, fill=Total_Deaths, 
                         size=Magnitude, label=City))

ggplot2::ggplot(data = eq_data) +
 geom_timeline_label(aes(x=Date, y=Country, fill=Total_Deaths, 
                         size=Magnitude, label=City), n_max=1)

}
