% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Create timeline label geom function}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "timelineLabel",
  position = "identity",
  na.rm = TRUE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Defaults to NULL}

\item{data}{Defaults to NULL, so the data is inherited from the plot
data as specified in the \code{ggplot2::ggplot()} call}

\item{stat}{Defaults to "timelineLabel", specifying that the
statistical transformation to be used on the data for this layer is
the timelineLabel stat}

\item{position}{Defaults to "identity"}

\item{na.rm}{Defaults to TRUE}

\item{show.legend}{Defaults to TRUE}

\item{inherit.aes}{Defaults to TRUE}
}
\value{
a layer which plots the data as per the GeomTimelineLabel
object definition that can be added to a ggplot object
}
\description{
This creates the \code{geom_timeline_label()} function.
It uses the ggproto object GeomTimelineLabel
This function allows the timeline and points (as specified in the
creation of the object GeomTimelineLabel) to be plotted as a ggplot
layer.
}
\section{Aesthetics}{

\code{geom_timeline_label()} understands the following aesthetics:
\itemize{
\item{x}{REQUIRED. A column of dates. Column type must be Date}
\item{xmin}{REQUIRED. Minimum/starting date. Must be a Date object}
\item{xmax}{REQUIRED. Maxmimum/ending date. Must be a Date object}
\item{txt}{REQUIRED. Any type of column to be displayed as text}
\item{max_by}{REQUIRED. A numeric/Date column to select maximum
observations by}
\item{n_max}{optional, if not supplied will be auto calculated.
A number specifying how many values should be selected from the
max_by column (arranged in descending order) }
\item{colour}{optional, defaults to black. Column to color points
by}
\item{fill}{optional, defaults to black. Column to fill points by}
\item{size}{optional, defaults to 5pt. Column to size points by}
\item{alpha}{optional, defaults to 1 (no transparency)}
}
}

\examples{
  \dontrun{ggplot2::ggplot(data=
                                eq_raw_cleaner(tail(eq,100)),
                           ggplot2::aes(x=Date,
                                        xmin=as.Date("2022-01-14"),
                                        xmax=as.Date("2022-05-26"),
                                        max_by=Mag,
                                        txt=Country,
                                        n_max=3)) +
                           geom_timeline_label()}

}
