% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline
This geom is intended to be used in conjunction with the geom_timeline geom
to add a vertical line with a text annotation}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data}{The earthquake data to be plotted. See the example for details.}
}
\value{
ggplot2 graphical object of earthquakes
}
\description{
geom_timeline
This geom is intended to be used in conjunction with the geom_timeline geom
to add a vertical line with a text annotation
}
\examples{
library(ggplot2)
eq = eq_clean_data("earthquakes-2020-10-19_15-21-05_+0300.tsv")
eq = dplyr::filter(eq, COUNTRY == "Mexico" & lubridate::year(DATE) >= 2000)
ggplot2::ggplot(eq,aes(x=DATE,y = COUNTRY, color = TOTAL_DEATHS, size = TOTAL_DEATHS)) +
  geom_timeline(alpha=.5) +
  geom_timeline_label(aes(label = LOCATION_NAME, n_max = 5))
}
