% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{print("GeomTimeline")
This is the function that actually does the job behind a call to the ggplot2
geom_timeline.}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimeline
}
\description{
print("GeomTimeline")
This is the function that actually does the job behind a call to the ggplot2
geom_timeline.
}
\keyword{datasets}
