% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
A Geom Object that will create a points Grob of a timeline.
}
\note{
x and y aesthetics are required. If you do not desire to use any variable as the y value, just set it to a constant value.
}
\keyword{datasets}
