% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomTimeLine.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom GeomTimeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
GeomTimeline
}
\description{
This is a geom for plotting a time line of earthquakes ranging from xmin to xmaxdates
with a point for each earthquake. The min and max dates are defined by filtering the dataset on the DATE field
Optional aesthetics include color, size, and alpha (for transparency).
The xaesthetic is a date and an optional y aesthetic is a factor indicating some stratification in which
case multiple time lines will be plotted for each level of the factor (e.g. country)
case multiple time lines will be plotted for each level of the factor (e.g. country)
}
\keyword{datasets}
