% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{print("GeomTimelineLabel")
This is the function that actually does the job behind a call to the ggplot2
geom_timeline_label.}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimelineLabel
}
\description{
print("GeomTimelineLabel")
This is the function that actually does the job behind a call to the ggplot2
geom_timeline_label.
}
\keyword{datasets}
