% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTimeline_Label.R
\docType{data}
\name{GeomTimeLineLabel}
\alias{GeomTimeLineLabel}
\title{GeomTimeLineLable}
\format{An object of class \code{GeomTimeLineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeLineLabel
}
\arguments{
\item{x}{A required aesthetic of representing the year.}

\item{label}{A required aesthetic of name of the location}
}
\description{
This is the GeomTimeLineLable. It creates the labels for the top n earthquakes
}
\keyword{datasets}
