% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timelinelabel.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Timelinelabel Geom}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
Timelinelabel Geom
}
\keyword{datasets}
