% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\alias{GeomTimeline}
\alias{StatTimelineLabel}
\alias{GeomTimelineLabel}
\title{Timeline Stat Object}
\description{
Timeline Stat Object

Timeline Geom Object

Labeled Timeline Stat Object

Labeled Timeline Geom Object
}
\keyword{datasets}
