% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_visualization.R
\name{stat_timeline}
\alias{stat_timeline}
\title{Create timeline stat function}
\usage{
stat_timeline(
  mapping = NULL,
  data = NULL,
  geom = "timeline",
  position = "identity",
  show.legend = TRUE,
  inherit.aes = TRUE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Defaults to NULL}

\item{data}{Defaults to NULL, so the data is inherited from the plot
data as specified in the \code{ggplot2::ggplot()} call}

\item{geom}{Defaults to "timeline"}

\item{position}{Defaults to "identity"}

\item{show.legend}{Defaults to TRUE}

\item{inherit.aes}{Defaults to TRUE}

\item{na.rm}{Defaults to FALSE}

\item{...}{Arguments to be passed to methods}
}
\value{
a layer which renders the data to make it suitable for the
timeline geom to plot
}
\description{
This creates the \code{stat_timeline()} function.
It uses the ggproto object StatTimeline.
The \code{stat_timeline} function takes data of a certain format,
processes it, and returns it in the correct format for the timeline
geom to plot.
}
\section{Aesthetics}{

\code{stat_timeline()} understands the following aesthetics:
\itemize{
\item{x}{REQUIRED. A column of dates. Column type must be Date}
\item{xmin}{REQUIRED. Minimum/starting date. Must be a Date object}
\item{xmax}{REQUIRED. Maxmimum/ending date. Must be a Date object}
}
}

\examples{
  \dontrun{ggplot2::ggplot(data=
                                eq_raw_cleaner(tail(eq,100)),
                           ggplot2::aes(x=Date,
                                        xmin=as.Date("2022-01-14"),
                                        xmax=as.Date("2022-05-26"))) +
                                        stat_timeline(size=5)}

}
