% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{theme_timeline}
\alias{theme_timeline}
\title{Timeline theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 9.}
\usage{
theme_timeline
}
\description{
New theme for earthquake  timeline
}
\keyword{datasets}
